last_update

+++ src/main/java/com/mycode/pathpilotserver/vehicles/models/Vehicle.java
@@ -105,7 +105,7 @@ public class Vehicle {
     }
 
     public void increaseKilometers(double kilometers) {
-        this.km += kilometers;
+        this.km += (int) kilometers;
     }
 
     @Override
